working with redis db, hono, and fetch cashu-ts
