+++ proxnut-proxy/package.json
@@ -12,9 +12,10 @@
     "typescript": "^5.0.0"
   },
   "dependencies": {
-    "@cashu/cashu-ts": "^0.8.0-rc.4",
-    "@elysiajs/cors": "^0.5.0",
-    "elysia": "^0.5.12",
+    "@gandlaf21/cashu-ts": "^0.8.0-rc.8",
+    "@elysiajs/cors": "^0.6.0-rc.0",
+    "elysia": "^0.6.0",
+    "hono": "^3.3.4",
     "redis": "^4.6.7"
   }
 }
