+++ proxnut-proxy/redis.ts
@@ -6,7 +6,7 @@ let client: undefined | RedisClientType;
 
 export const getRedisInstance = async () => {
   if (!client) {
-    client = createClient({ socket: { host: "localhost" } });
+    client = createClient({ socket: { host: "redis" } });
     client.on("error", (err) => console.log("Redis Client Error", err));
     await client.connect();
   }
